# Significantly improve performance of functions that check IsLANServer()