@peace-maker 2025-02-10
How do static initializers work again? I'd expect the change to be a no-op since the global variable isn't used elsewhere.

Shouldn't you need something like this? I'm confused by your benchmarking.

```cpp
static ConVar* sv_lan = nullptr;
if (!sv_lan)
  sv_lan = FindVar(...);
if (!sv_lan)
   // Error
```