@Kenzzer 2025-02-10
> How do static initializers work again?

According to the C++ standard. (Taken from https://isocpp.org/files/papers/N4860.pdf)

> Dynamic initialization of a block-scope variable with static storage duration (6.7.5.1) or thread storage
> duration (6.7.5.2) is performed the first time control passes through its declaration; such a variable is
> considered initialized upon the completion of its initialization. If the initialization exits by throwing an
> exception, the initialization is not complete, so it will be tried again the next time control enters the declaration.
> If control enters the declaration concurrently while the variable is being initialized, the concurrent execution
> shall wait for completion of the initialization.85 If control re-enters the declaration recursively while the
> variable is being initialized, the behavior is undefined.

So this seems fine to keep as is, the behaviour is well-defined.

> I'm confused by your benchmarking.

I'm also confused, but if I'm deciphering their wording properly. They've optimized `GetClientAuthId` with `AuthId_Steam2` to run at an average of `~0.000044 ms` (new) vs `0.0401949994 ms` (previously).
> `GetClientName is 52.38%, 2.09x faster than AuthId_Steam2`

`GetClientName avg: 0.0000220000 ms` times two (for two times slower), would be `~0.000044 ms`. Though yeah posting the numbers directly would have been more meaningful.